increase the maximum number of nested prototypes

+++ src/inc/any/asm.h
@@ -55,10 +55,10 @@ ANY_API astring_ref_t any_asm_string_to_ref(aasm_t* self, const char* s);
 // Extend prototype with more capacity.
 ANY_API void any_asm_grow(
     aasm_t* self,
-    uint32_t max_instructions,
+    int32_t max_instructions,
     uint8_t max_constants,
     uint8_t max_imports,
-    uint8_t max_nesteds);
+    int16_t max_nesteds);
 
 // Get prototype header.
 ANY_API aasm_prototype_t* any_asm_prototype(aasm_t* self);
